fix cover upload, not deleted default cover image

+++ app/Http/Controllers/Admin/BookController.php
@@ -22,7 +22,7 @@ class BookController extends Controller
         ->with('books', 
         Book::with('authors', 'genres', 'reviews')
         ->latest()
-        ->paginate(10));
+        ->paginate(20));
     }
 
     /**
@@ -102,11 +102,13 @@ class BookController extends Controller
                     'cover' => 'required|image|mimes:jpeg,png,jpg,gif,svg|max:2048'
                 ]);
 
-
-                if(Storage::exists('public/'.$book->cover)){
-                    Storage::delete('public/'.$book->cover);
-                }else{      
-                    dd('File does not exists.');
+                 //kad neissitrintu default cover
+                if($book->cover != 'cover.png'){
+                    if(Storage::exists('public/'.$book->cover)){
+                        Storage::delete('public/'.$book->cover);
+                    }else{      
+                        dd('File does not exists.');
+                    }
                 }
 
 
@@ -166,8 +168,12 @@ class BookController extends Controller
     {
         $book->genres()->detach();
         $book->authors()->detach();
+
+        $book->reviews()->delete();
         
         $book->delete();
+
+        //$book->reports->delete();
         
         //gristam i pradini puslapi
         //siunciam pranesima kad irasymas atliktas
